// src/app/api/model/cu-customer-model.ts
/**
 * CrediFlow.API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface CUCustomerModel { 
    customerId?: string | null;
    nationalId?: string | null;
    customerCode?: string | null;
    fullName?: string | null;
    dateOfBirth?: string | null;
    gender?: string | null;
    phone?: string | null;
    address?: string | null;
    firstSourceType?: string | null;
    hasBadHistory?: boolean;
    badHistoryNote?: string | null;
    firstStoreId?: string | null;
    referredByCollaboratorId?: string | null;
}